[internal] - Change to using sobek instead of goja
